+++ RustyBags/VersionHandshake.cs
@@ -1,9 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Reflection;
-using System.Security.Cryptography;
-using System.Text;
 using HarmonyLib;
 
 namespace RustyBags