+++ RustyBags/src/BagInventory.cs
@@ -1,30 +1,78 @@
-﻿using HarmonyLib;
+﻿using System.Linq;
+using HarmonyLib;
 using JetBrains.Annotations;
+using RustyBags.Managers;
 using RustyBags.Utilities;
 using UnityEngine;
 
 namespace RustyBags;
 
+public static class InventoryExtensions
+{
+    public static bool HasBag(this Inventory inventory) => inventory.m_inventory.FirstOrDefault(x => x is Bag) != null;
+    public static bool IsPlayerInventory(this Inventory inventory) => Player.m_localPlayer && inventory == Player.m_localPlayer.GetInventory();
+}
+
+
 [HarmonyPatch(typeof(InventoryGrid), nameof(InventoryGrid.DropItem))]
 public static class InventoryGrid_DropItem_Patch
 {
     [UsedImplicitly]
-    private static bool Prefix(InventoryGrid __instance, Inventory fromInventory, ItemDrop.ItemData item, Vector2i pos)
+    private static bool Prefix(InventoryGrid __instance, Inventory fromInventory, ItemDrop.ItemData item, Vector2i pos, ref bool __result)
     {
-        return __instance.m_inventory is not BagInventory bagInventory || bagInventory.CanAddItem(item);
+        if (fromInventory == __instance.m_inventory) return true;
+        
+        if (!Configs.MultipleBags && __instance.m_inventory.IsPlayerInventory() && __instance.m_inventory.HasBag())
+        {
+            __result = false;
+            return false;
+        }
+    
+        if (__instance.m_inventory is BagInventory bagInventory && !bagInventory.CanAddItem(item))
+        {
+            __result = false;
+            return false;
+        }
+    
+        return true;
     }
+    
 }
 
 [HarmonyPatch(typeof(Inventory), nameof(Inventory.AddItem), typeof(ItemDrop.ItemData))]
 public static class Inventory_AddItem_Patch
 {
+    [HarmonyPriority(Priority.First)]
     [UsedImplicitly]
     private static bool Prefix(Inventory __instance, ItemDrop.ItemData item, ref bool __result)
     {
-        if (__instance is not BagInventory bagInventory || bagInventory.CanAddItem(item)) return true;
-        __result = false;
-        return false;
+        if (!Configs.MultipleBags && item is Bag && __instance.IsPlayerInventory() && __instance.HasBag())
+        {
+            __result = false;
+            return false;
+        }
+
+        
+        if (__instance is BagInventory bagInventory && !bagInventory.CanAddItem(item))
+        {
+            __result = false;
+            return false;
+        }
+        
+        return true;
     }
+}
+
+[HarmonyPatch(typeof(Inventory), nameof(Inventory.CanAddItem), typeof(ItemDrop.ItemData),typeof(int))]
+public static class Inventory_CanAddItem_Patch
+{
+    [UsedImplicitly]
+    private static void Postfix(Inventory __instance, ItemDrop.ItemData item, ref bool __result)
+    {
+        if (Configs.MultipleBags || !__result || item is not Bag || !__instance.IsPlayerInventory() || !__instance.HasBag()) return;
+        __result = false;
+    }
+
 }
 
 public class BagInventory : Inventory